Lists separated into a way that I prefer.
Mainly the separation is the social media, so that I can enable access to one social media site at a time or needed.

Example:
- All social media sites blocked.
- To access Facebook, I enable access by disable the DNS blocking for Facebook, while still keeping the Trackers list enabled. In this way I just like that while using a social media site, no other social media site is being communicated to (because they like to sleep in the same bed with each other).
- When done on Facebook, I can turn on blocking for Facebook, and then go to a different social media site as needed, repeating the same principal.

The other lists were separated in a way that they can target a category more specifically, for someone else that may prefer it that way.

But as for me, I don't have need to disable the other lists, Trackers, Gambling, Faceknews, Adults sites, etc. (if to come).